Solve 90 degree rotation for skinned meshes.
The post import script was folding rotations into the root when the scene had skinned meshes at root but preserve hierarchy is disabled.
Godot makes no distinction as to where in the glTF hierarchy a skinned mesh is, since it is always reparented to the Skeleton3D.
However, Unity takes its hierarchy into account when determining whether or not to remove the root node when Preserve Hierarchy is disabled.
This had caused the 90 degree rotation to be attached to the wrong node, and scenes would override it incorrectly.

Fixes #10